separated screens to individual files added strings to resources and created mvvm packages
