--- app/src/main/java/com/demo/jetpackcomposecrashcourse/ContentScreens.kt
@@ -1,161 +1,0 @@
-package com.demo.jetpackcomposecrashcourse
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Divider
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
-import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Edit
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-
-
-@Composable
-fun HomeScreen() {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(colorResource(id = R.color.white))
-    ) {
-        Text(
-            text = "Welcome back!",
-            fontWeight = FontWeight.ExtraBold,
-            color = Color.Black,
-            textAlign = TextAlign.Right,
-            fontSize = 24.sp,
-            modifier = Modifier
-                .padding(top = 52.dp, start = 15.dp)
-        )
-        Image(
-            painter = painterResource(id = R.drawable.medicine),
-            contentDescription = "Medicine Art",
-            modifier = Modifier
-                .size(300.dp)
-                .align(alignment = Alignment.CenterHorizontally)
-                .padding(top = 10.dp)
-        )
-        Text(
-            text = "Schedule",
-            fontWeight = FontWeight.ExtraBold,
-            color = Color.Black,
-            textAlign = TextAlign.Right,
-            fontSize = 24.sp,
-            modifier = Modifier
-                .padding(top = 15.dp, start = 15.dp)
-        )
-        RowList(rows = listOf("Joshua", "Jermain", "Kendrick"))
-    }
-}
-
-@Composable
-fun MedicationScreen() {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(colorResource(id = R.color.white))
-            .wrapContentSize(Alignment.Center)
-    ) {
-        Text(
-            text = "Medication View",
-            fontWeight = FontWeight.Bold,
-            color = Color.Black,
-            modifier = Modifier.align(Alignment.CenterHorizontally),
-            textAlign = TextAlign.Center,
-            fontSize = 25.sp
-        )
-    }
-}
-
-@Composable
-fun MetricsScreen() {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(colorResource(id = R.color.white))
-            .wrapContentSize(Alignment.Center)
-    ) {
-        Text(
-            text = "Metrics View",
-            fontWeight = FontWeight.Bold,
-            color = Color.Black,
-            modifier = Modifier.align(Alignment.CenterHorizontally),
-            textAlign = TextAlign.Center,
-            fontSize = 25.sp
-        )
-    }
-}
-
-@Composable
-fun UserProfileScreen() {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(colorResource(id = R.color.white))
-            .wrapContentSize(Alignment.Center)
-    ) {
-        Text(
-            text = "User Profile View",
-            fontWeight = FontWeight.Bold,
-            color = Color.Black,
-            modifier = Modifier.align(Alignment.CenterHorizontally),
-            textAlign = TextAlign.Center,
-            fontSize = 25.sp
-        )
-    }
-}
-
-@Composable
-fun RowList(rows: List<String>) {
-    LazyColumn(
-        modifier = Modifier.fillMaxWidth(),
-        contentPadding = PaddingValues(horizontal = 38.dp, vertical = 29.dp)
-    ) {
-        items(rows.size) { index ->
-            RowItem(rows[index])
-            if (index < rows.size - 1) {
-                Divider(thickness = 14.dp, color = colorResource(id = R.color.white))
-            }
-        }
-    }
-}
-
-@Composable
-fun RowItem(text: String) {
-
-    Row(
-        modifier = Modifier
-            .clip(RoundedCornerShape(5.dp))
-            .fillMaxWidth()
-            .height(height = 81.dp)
-            .background(color = colorResource(id = R.color.light_blue)),
-        horizontalArrangement = Arrangement.SpaceEvenly,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Text(text = text)
-        IconButton(onClick = { /* Do something */ }) {
-            Icon(Icons.Filled.Edit, contentDescription = "Favorite")
-        }
-    }
-}
-
-
-@Preview(showBackground = true)
-@Composable
-fun HomeScreenPreview() {
-    HomeScreen()
-}
