+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,10 @@
 <resources>
     <string name="app_name">jetpackcomposecrashcourse</string>
+    <string name="welcome">Welcome back!</string>
+    <string name="schedule">Schedule</string>
+    <string name="home_view">Home View</string>
+    <string name="med_view">Medication View</string>
+    <string name="metrics_view">Metrics View</string>
+    <string name="user_view">User Profile View</string>
+    <string name="medicine_art">Medicine Art</string>
 </resources>
